PB-1080: show both warning and error feedback
- Issue : When we have both an error feedback and a warning feedback, they would be displayed one atop each other rather than in a column.
- Fix : We create a small component to hold both the errors and warning feedback.
